Split DMA peripheral mappings into processor families
